feat(admin-gui): makes sure token exists before attempting an API request and adds some html hide/show the project stuff depending on the login status

+++ services/admin/frontend/Projects.svelte
@@ -6,9 +6,10 @@
     import Modal from 'svelte-simple-modal';
 
     onMount(() => {
-        currentUser.subscribe(async info => {
-            console.log("Projects currentUser: ", $currentUser)
-            await getProjects(info.token);
+        currentUser.subscribe(async userInfo => {
+            if ($currentUser.token) {
+                await getProjects(userInfo.token);
+            }
         });
     });
 
@@ -19,26 +20,32 @@
     <div>
         <h1>Projects</h1>
     </div>
-    <div class="list">
-        {#each $projectsList as p}
-           <li>
-               <div class="name">
-                   <Router>
-                       <Link to={`/projects/${p.id}`} let:params>
-                           {p.longName}
-                       </Link>
-                   </Router>
-               </div>
-               <div class="delete">
-                   <button on:click={deleteProject($currentUser.token, p.id)}>X</button>
-               </div>
-           </li>
-        {/each}
-        <!--    Modal for creating a new project-->
-        <Modal>
-            <Content modalType="create" token="{$currentUser.token}"/>
-        </Modal>
-    </div>
+    {#if $currentUser.token}
+        <div class="list">
+            {#each $projectsList as p}
+               <li>
+                   <div class="name">
+                       <Router>
+                           <Link to={`/projects/${p.id}`} let:params>
+                               {p.longName}
+                           </Link>
+                       </Router>
+                   </div>
+                   <div class="delete">
+                       <button on:click={deleteProject($currentUser.token, p.id)}>X</button>
+                   </div>
+               </li>
+            {/each}
+            <!--    Modal for creating a new project-->
+            <Modal>
+                <Content modalType="create" token="{$currentUser.token}"/>
+            </Modal>
+        </div>
+    {:else}
+        <div>
+            <p>You must be logged in to access the list of projects.</p>
+        </div>
+    {/if}
 </div>
